+++ db/migrate/20200331213717_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
       t.decimal :totalamount
-      t.int :quantity
+      t.decimal :quantity
 
       t.timestamps
     end
